+++ src/blockchain.rs
@@ -18,8 +18,9 @@ use tokio::{
   task::JoinHandle,
 };
 use tracing::{debug, error, info};
+use tracing_futures::Instrument;
 
-type AccountsMap = DashMap<Pubkey, Account>;
+pub(crate) type AccountsMap = DashMap<Pubkey, (Account, bool)>;
 pub(crate) type SubRequestCall =
   (SubRequest, Option<oneshot::Sender<Vec<Account>>>);
 
@@ -30,6 +31,7 @@ pub(crate) type SubRequestCall =
 /// for now it is set to 64, which gives us about 30 seconds on Solana
 /// as there can be an update at most once every 400 miliseconds (blocktime)
 const MAX_UPDATES_SUBSCRIBER_LAG: usize = 64;
+const MIN_RECONNECT_EVERY: u64 = 30;
 
 #[derive(Clone)]
 pub struct SyncOptions {
@@ -73,6 +75,16 @@ pub struct BlockchainShadow {
 impl BlockchainShadow {
   pub async fn new(options: SyncOptions) -> Result<Self> {
     let max_lag = options.max_lag.unwrap_or(MAX_UPDATES_SUBSCRIBER_LAG);
+
+    // protect our enduser against to quick reconnects which would
+    // cause unstable behaviour
+    if options
+      .reconnect_every
+      .map(|v| v < Duration::from_secs(MIN_RECONNECT_EVERY))
+      .unwrap_or(false)
+    {
+      panic!("low reconnect_every duration causes unstable behaviour, minimum reconnect_every value is {} seconds", MIN_RECONNECT_EVERY)
+    }
 
     let (subscribe_tx, subscribe_rx) = unbounded_channel::<SubRequestCall>();
 
@@ -167,13 +179,13 @@ impl BlockchainShadow {
   pub fn for_each_account(&self, mut op: impl FnMut(&Pubkey, &Account)) {
     for pair in self.accounts.iter() {
       let pubkey = pair.pair().0;
-      let account = pair.pair().1;
+      let (account, _) = pair.pair().1;
       op(pubkey, account);
     }
   }
 
   pub fn get_account(&self, key: &Pubkey) -> Option<Account> {
-    self.accounts.get(key).map(|acc| acc.clone())
+    self.accounts.get(key).map(|acc| acc.clone().0)
   }
 
   pub async fn worker(mut self) -> Result<()> {
@@ -202,39 +214,49 @@ impl BlockchainShadow {
       self.options.rpc_timeout,
       self.options.commitment,
     );
-    self.sync_worker = Some(tokio::spawn(async move {
-      let mut listener =
-        SolanaChangeListener::new(client, accs_ref.clone(), options).await?;
-      loop {
-        tokio::select! {
-          recv_result = listener.recv() => {
-            match recv_result {
-              Ok(Some(AccountUpdate { pubkey, account })) => {
-                debug!("account {} updated", &pubkey);
-                accs_ref.insert(pubkey, account.clone());
-                if updates_tx.receiver_count() != 0 {
-                  updates_tx.send((pubkey, account)).unwrap();
+    self.sync_worker = Some(tokio::spawn(
+      async move {
+        let mut listener =
+          SolanaChangeListener::new(client, accs_ref.clone(), options).await?;
+        loop {
+          tokio::select! {
+            recv_result = listener.recv() => {
+              match recv_result {
+                Ok(Some(AccountUpdate { pubkey, account })) => {
+                  debug!("account {} updated", &pubkey);
+                  accs_ref.insert(pubkey, ( account.clone(), true ));
+                  if updates_tx.receiver_count() != 0 {
+                    updates_tx.send((pubkey, account)).unwrap();
+                  }
+                },
+                Ok(None) => {
+                  unreachable!();
+                },
+                Err(e) => {
+                  error!("error in the sync worker thread: {:?}", e);
+                  listener.reconnect_all().await?;
                 }
-              },
-              Ok(None) => {
-                unreachable!();
-              },
-              Err(e) => {
-                error!("error in the sync worker thread: {:?}", e);
-                listener.reconnect_all().await?;
+              }
+            },
+            Some(subreq) = subscribe_rx.recv() => {
+              match subreq {
+                ( SubRequest::Account(pubkey), oneshot ) => {
+                    debug!(?pubkey, "subscribe_account recv");
+                    listener.subscribe_account(pubkey, oneshot).await? },
+                ( SubRequest::Program(pubkey), oneshot ) => {
+                    debug!(?pubkey, "subscribe_program recv");
+                    listener.subscribe_program(pubkey, oneshot).await? },
+                ( SubRequest::ReconnectAll, _ ) => {
+                    debug!("reconnect_all recv");
+                    listener.reconnect_all().await?
+                }
               }
             }
-          },
-          Some(subreq) = subscribe_rx.recv() => {
-            match subreq {
-              ( SubRequest::Account(pubkey), oneshot ) => listener.subscribe_account(pubkey, oneshot).await?,
-              ( SubRequest::Program(pubkey), oneshot ) => listener.subscribe_program(pubkey, oneshot).await?,
-              ( SubRequest::ReconnectAll, _ ) => listener.reconnect_all().await?
-            }
-          }
-        };
+          };
+        }
       }
-    }));
+      .instrument(tracing::debug_span!("worker_loop")),
+    ));
 
     Ok(())
   }
@@ -242,15 +264,18 @@ impl BlockchainShadow {
   async fn create_monitor_worker(&mut self) -> Result<()> {
     if let Some(every) = self.options.reconnect_every {
       let channel = self.sub_req.clone();
-      self.monitor_worker = Some(tokio::spawn(async move {
-        loop {
-          info!("reestablising connection to solana");
-          tokio::time::sleep(every).await;
-          channel
-            .send((SubRequest::ReconnectAll, None))
-            .map_err(|_| Error::InternalError)?;
+      self.monitor_worker = Some(tokio::spawn(
+        async move {
+          loop {
+            tokio::time::sleep(every).await;
+            info!("reestablising connection to solana");
+            if let Err(e) = channel.send((SubRequest::ReconnectAll, None)) {
+              tracing::error!(?e)
+            }
+          }
         }
-      }));
+        .instrument(tracing::debug_span!("monitor_worker_loop")),
+      ));
     }
     Ok(())
   }