+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.2.1
+- tracking of change on accounts for reconnects
+- panic! on too low reconnect values, to avoid unstable behaviour
+- instrumentation, tracing_futures RUST_LOG=solana_shadow=debug
+
 # 0.2.0
 
 - API change, `add_program`, `add_accounts` and `add_account` now return accounts